+++ backup_codes/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class BackupCodesConfig(AppConfig):
+    name = "backup_codes"
