+++ backup_codes/templates/backup_codes/backup_codes_list.html
@@ -10,10 +10,10 @@
 
   <div class="static-codes">
     <ul>
-    {% for security_code in security_code_list %}
+    {% for backup_code in backup_code_list %}
       <li>
-        <span class="visually-hidden">{{ security_code.token|upper|join:" " }}</span>
-        <span aria-hidden="true"><span>{{ security_code.token|slice:":4" }}</span><span>{{ security_code.token|slice:"4:" }}</span></span>
+        <span class="visually-hidden">{{ backup_code.token|upper|join:" " }}</span>
+        <span aria-hidden="true"><span>{{ backup_code.token|slice:":4" }}</span><span>{{ backup_code.token|slice:"4:" }}</span></span>
       </li>
     {% endfor %}
     </ul>
