+++ backup_codes/urls.py
@@ -1,0 +1,12 @@
+from django.urls import path
+
+from . import views
+
+
+urlpatterns = [
+    path(
+        "backup-codes",
+        views.BackupCodeListView.as_view(),
+        name="backup_codes",
+    )
+]
