+++ portal/settings.py
@@ -83,6 +83,7 @@ INSTALLED_APPS = [
     "covid_key",
     "invitations",
     "contact",
+    "backup_codes",
     "axes",
     "django.contrib.sites",  # Required for invitations
     "django.contrib.auth",
