+++ portal/urls.py
@@ -62,6 +62,7 @@ urlpatterns += i18n_patterns(
     path("", include("profiles.urls")),
     path("", include("covid_key.urls")),
     path("contact/", include("contact.urls")),
+    path("", include("backup_codes.urls")),
     path(
         "invitations/",
         include(invitation_patterns, namespace="invitations"),
