+++ profiles/static/scss/styles.scss
@@ -6,7 +6,7 @@
 @import "./_base";
 
 /* Page-specific styles */
-@import "./pages/_2fa-codes";
+@import "./pages/_backup-codes";
 @import "./pages/_login";
 @import "./pages/_number";
