+++ profiles/templates/profiles/healthcareuser_detail.html
@@ -94,13 +94,13 @@
                   </a>
               </td>
             </tr>
-            {% switch "SECURITY_CODE" %}
+            {% switch "BACKUP_CODE" %}
             <tr>
               <th scope="row">{% trans "Security Codes" %}</th>
               <td>{{ security_code_count }} {%trans " unused codes remaining" %}</td>
               {% if not view_only %}
               <td>
-                  <a href="{% url 'security_code_list' %}">
+                  <a href="{% url 'backup_codes' %}">
                     <span class="visually-hidden">
 
                         {% trans "Get Security Codes" %}
