+++ profiles/tests.py
@@ -7,7 +7,6 @@ from django.contrib.auth import get_user_model
 from django.utils import translation, timezone
 from django.core.exceptions import ValidationError
 from django_otp import DEVICE_ID_SESSION_KEY
-from django_otp.plugins.otp_static.models import StaticDevice
 from invitations.models import Invitation
 from axes.models import AccessAttempt
 from datetime import datetime, timedelta
@@ -114,7 +113,7 @@ class AdminUserTestCase(TestCase):
         self.invited_email = "invited@test.com"
         AuthorizedDomain.objects.create(domain="test.com")
 
-        Switch.objects.create(name="SECURITY_CODE", active=True)
+        Switch.objects.create(name="BACKUP_CODE", active=True)
 
     def login(self, credentials: dict = None, login_2fa: bool = True):
         if credentials is None:
@@ -1061,13 +1060,12 @@ class ProfileView(AdminUserTestCase):
         self.assertEqual(response.status_code, 403)
 
     def test_only_user_can_view_own_security_codes(self):
-
         self.login()
         user2 = User.objects.create_user(**get_other_credentials(is_admin=True))
         response = self.client.get(reverse("user_profile", kwargs={"pk": user2.id}))
         self.assertEqual(response.status_code, 200)
         ## no security codes link present
-        self.assertNotContains(response, '<a href="/en/profiles/security-codes">')
+        self.assertNotContains(response, '<a href="/en/backup-codes">')
 
 
 class DeleteView(AdminUserTestCase):
@@ -1271,67 +1269,3 @@ class ProfileEditView(AdminUserTestCase):
         self.assertEqual(response.status_code, 302)
         user = HealthcareUser.objects.get(pk=self.user.id)
         self.assertEqual(user.phone_number, number)
-
-
-class SecurityCodeView(AdminUserTestCase):
-    def setUp(self):
-        super().setUp(is_admin=True)
-
-    def test_user_can_view_security_codes_row_on_account_page(self):
-        self.login()
-        response = self.client.get(reverse("user_profile", kwargs={"pk": self.user.id}))
-        self.assertEqual(response.status_code, 200)
-        ## security codes link is present
-        self.assertContains(response, '<a href="/en/profiles/security-codes">')
-
-    def test_user_can_generate_5_security_codes(self):
-        self.login()
-        response = self.client.get(reverse("security_code_list"))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "<h1>Security Codes</h1>")
-
-        device = StaticDevice.objects.get(user__id=self.user.id)
-        self.assertEqual(len(device.token_set.all()), 5)
-
-    def test_old_security_codes_invalid_if_new_security_codes_generated(self):
-        self.login()
-        response = self.client.get(reverse("security_code_list"))
-        self.assertEqual(response.status_code, 200)
-
-        old_codes = [
-            t.token
-            for t in StaticDevice.objects.get(user__id=self.user.id).token_set.all()
-        ]
-
-        response = self.client.get(reverse("security_code_list"))
-        self.assertEqual(response.status_code, 200)
-        new_codes = [
-            t.token
-            for t in StaticDevice.objects.get(user__id=self.user.id).token_set.all()
-        ]
-
-        for code in old_codes:
-            self.assertNotIn(code, new_codes)
-
-    def test_user_codes_HTML(self):
-        self.login()
-        response = self.client.get(reverse("security_code_list"))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, "<h1>Security Codes</h1>")
-        device = StaticDevice.objects.get(user__id=self.user.id)
-
-        tokens = [t.token for t in device.token_set.all()]
-
-        for token in tokens:
-            self.assertContains(
-                response,
-                '<span class="visually-hidden">{}</span>'.format(
-                    " ".join(token.upper())
-                ),
-            )
-            self.assertContains(
-                response,
-                '<span aria-hidden="true"><span>{}</span><span>{}</span>'.format(
-                    token[:4], token[-4:]
-                ),
-            )
